+++ src/raid/raid1/tests/misc/replica_states.cpp
@@ -86,19 +86,19 @@ static void expect_swap_a_success(std::shared_ptr< ublkpp::TestDisk >& new_devic
             memset(iovecs->iov_base, 0x00, iovecs->iov_len);
             return ublkpp::raid1::k_page_size;
         });
-    EXPECT_CALL(*new_device, sync_iov(UBLK_IO_OP_WRITE, _, _, _))
-        .Times(2)
-        .WillOnce([&raid_device](uint8_t, iovec* iovecs, uint32_t nr_vecs, off_t addr) -> io_result {
-            // init_to: bitmap write to new device
-            EXPECT_EQ(1U, nr_vecs);
-            EXPECT_EQ(ublkpp::raid1::k_page_size, ublkpp::iovec_len(iovecs, iovecs + nr_vecs));
+    // init_to writes _num_pages zero pages to new device (capacity-derived, batched)
+    EXPECT_CALL(*new_device, sync_iov(UBLK_IO_OP_WRITE, _, _, testing::Gt((off_t)0)))
+        .Times(testing::AtLeast(1))
+        .WillRepeatedly([&raid_device](uint8_t, iovec* iovecs, uint32_t nr_vecs, off_t addr) -> io_result {
             EXPECT_GE(addr, ublkpp::raid1::k_page_size);
             EXPECT_LT(addr, raid_device.reserved_size());
             EXPECT_EQ(0, isal_zero_detect(iovecs->iov_base, ublkpp::raid1::k_page_size)); // all zeros
-            return ublkpp::raid1::k_page_size;
-        })
+            return ublkpp::iovec_len(iovecs, iovecs + nr_vecs);
+        });
+    // __swap_device: new device superblock commit (new_device is in A slot → device_b=0)
+    EXPECT_CALL(*new_device, sync_iov(UBLK_IO_OP_WRITE, _, _, 0))
+        .Times(1)
         .WillOnce([](uint8_t, iovec* iovecs, uint32_t nr_vecs, off_t addr) -> io_result {
-            // __swap_device: new device superblock commit (new_device is in A slot → device_b=0)
             EXPECT_EQ(1U, nr_vecs);
             EXPECT_EQ(ublkpp::raid1::k_page_size, ublkpp::iovec_len(iovecs, iovecs + nr_vecs));
             EXPECT_EQ(0UL, addr);